// business/additional.d.ts
export type Products = {
    products:{}
}[]

export type Product = {
    // Id marked as optional due to its generated by mongo
    // we just need to check if when used the type is correct
    // however we don't used it when adding a new product
    _id?:number,
    title:string,
    desc:string,
    category:string,
    subcategory?:string,
    price:number,

}